Use variadic templates to replace swprintf. Microsoft's treatment of %s/%S made it non-portable

+++ Archs/ARM/CThumbInstruction.cpp
@@ -301,7 +301,7 @@ bool CThumbInstruction::Validate()
 
 		if (Vars.ImmediateBitLen != 32)
 		{
-			if (abs(Vars.Immediate) >= (unsigned int)(1 << Vars.ImmediateBitLen))
+			if (abs(Vars.Immediate) >= (1 << Vars.ImmediateBitLen))
 			{
 				Logger::queueError(Logger::Error,L"Immediate value %X out of range",Vars.Immediate);
 				return false;
